LayoutTests/http/tests/websocket/tests/hybi/websocket-cookie-overwrite-behavior-expected.txt

Tests WebSocket Set-Cookie overwriting behavior with respect to a document cookie.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Same origin WebSocket:
PASS cookieValue is "foo=bar"

Cross origin WebSocket:
PASS cookieValue is ""
PASS successfullyParsed is true

TEST COMPLETE
